Add a component test!
Assert some stuff about what should be in the component for different props

This invariant-oriented tooling an be extended later to generate a "props space" of
combinations of valid properties, and assert the invariants across those combinations.

It doesn't solve component testing -- we need to get actions in on this for that. But it's a
start.