Check components by throwing random valid props at them
and optionally assert that certain invariants about the output hold.

It's not the prettiest code, but simple enough and so far it works. we'll probably want to extend it to be able to specifically check certain props for regressions.

Not covered by this kind of component test: ensuring that components are passing valid props to their children. That's definitely a real, common error that can break the app, so we should figure out a way to test for it. To do that, we might need jsDOM though.