Typecheck nested components
The shallow dom-less render doesn't render child components, but it does return a tree with their component classes in place, which means we can actually see their proptypes and check the props that we're giving them! So, we can
ensure that a component is always providing legal proptypes to all of its children. I think we're really close to being set with this!

The biggest missing connection remaining for awesome client-side coverage is DOM events -- ensuring that we're creating valid action payloads. We will actually need jsDOM for this, but it's certainly possible.