Add component test invariant checking and import alias
Increase the power of component tests one step by allowing checkComponent to assert invariants of the output based on props!

This is mostly just re-adding previous work from `component-tests` so that I can turn back on the temporarily-disabled original component tests

The tests crash if the invariant description is missing to keep invariant specs well-documented and easy to trace when they're violated.

Imports to `test-utils` are now aliased so starting the path with `test-utils` get automatically resolved: Lower friction and hopefully bump dev happiness slightly when writing tests :)