Longer, formatted event descriptions
The text input was too short to include the full, properly-formatted programme for Angela's recital, and the public display lost all linebreak/para info. I tripled the max length to 1200, but this is still arbitrary...

I (finally!) created a file in `shared/` called `form-constraints.js`, to hold the shared 1200-char limit value, used by both the client and server.

This commit also fixes a bug in the text formatter where bad `key` props could cause subsequent paragraphs to vanish. Now `key` is provided correctly, and there are tests to assert this, woo!