Don't check proptype key coverage
With decorators, we can't do it effectively.

Also uncover a bunch of problems and fix everything:

- mocking console.error doesn't work because prop-types doesn't always log: https://stackoverflow.com/q/41916992/1299695
- shallow rendering isn't enough with decorators because shallow is just the decorator

So, now ratehub is the owner of a brand new shiny repo for better proptype checking, check-prop-types: https://github.com/ratehub/check-prop-types

And we now recursively shallow-render our way down the tree. So, deep rendering, I guess.