Add invariant checking hook
I'm going with a different API than we used in openride. This version is a little more agnostic about the rendering back-end used on the JSX that things are asserted against. The JSX is created, but rendering it to check is completely up to the user.

`invariant` (which is just the same-named npm package) is passed to the check callback, and that's how problems are signalled. So the usage of the word `invariant` is a little more cohesive than it was.

Skipping tests is left up to the user: just check props and return early.

Instead of having an array of checks with descriptions, there is just one check callback. Descriptions are coded into the invariants instead.