Recurse into components without proptypes
Before this change, certain render subtrees were skipped when a component didn't have proptypes. This was bad if the component under test was wrapped in eg. mobx-react's `inject()`: that wrapper doesn't have proptypes, so the component wouldn't be checked at all!

Now subtrees are always traversed even if the component doesn't have proptypes.